Avoid the need to pass the template id
I'll be using proper form builder helpers for the inputs of these forms
so that proper error messages can be displayed around each one. In order
to do that, each template must have different base input names to be
able to distinguish which question each answer option belongs to. So we
render a separate template next to each "add field" button, and we can
locate it from the input button itself without needing any ids.