Add rich text description to questions (#3066)
* Normalize file sections

Whitespace between attribute types.

* Remove `enable_tabs` unused parameter

* Remove unnecessary variable

* Allow import quillEditor creator from the outside

* Some whitespacing

* Fix typo

* Simplify survey specs dealing w/ translated fields

These specs needed to be this convoluted because previously we required
every locale to be filled in. Now we don't and things would get even
tricker because once I add a description field for the question we will
have several of this fields per question, so it gets really messy to
pick up the right selectors. Let's simplify this.

* Add a rich text description to questions