Fix bug in select field option filling
When adding new questions, their max number of options selector would be
prefilled with incorrect options because of two bugs:

* The main id of the dynamically added templete was not being replaced
with a unique value, so that was causing the number of options in the
select field to be messed up since when more than one dynamically fields
were added, they wouldn't have unique id.

* The count of the number of answer options in a question was incorrect
because it was using a global selector instead of one scope to the
particular question markup.