Allow impersonating regular users + big refactoring of impersonations (#3226)
* Normalize some controllers

Prefer fully qualified constants to `require_dependency`.

* Rename query to be more generic

* Use association in the model

* Fix typos

* Tiny spec rewording

* Allow reusing users query for other filterings

* Fix typo

* Fix bad HTML

* Spec wording fix

* Allow filtering managed users

* Offer all users for impersonation

* Typos

* Spec rewording

* Impersonations for everyone WIP

* Remove redundant `page.`

* Rename impersonation classes for the new situation

Where every user will be impersonable.

* Fix typo

* Rename shared example to match new situation

* Explicitly pass impersonated user to shared spec

* Tweak util method to accept a generic user

* Fix spec that was only passing by luck

The `unique_id` was being set wrong.

* Move impersonation logs to a separate controller

To prepare for further refactoring.

* Remove show and destroy managed user routes

Since the actions don't exist in the controller.

* Rename "Managed Users" section to "Impersonations"

And move the index action to the impersonations controller. This is part
of the work to reconcile things after allowing every user to be
impersonated.

* Normalize managed user abilities

* Only allow authorization handlers for impersonation

Authorization workflows are forbidden.

* Remove stuff we don't really need

We're now authorizing this via abilities.

* Normalize controllers

With the intention of later refactoring

* Pick authorization through the inner builder

* Merge impersonation flow to a single step

* Other dummy

* Merge impersonation forms into a single one

By adding an optional user attribute to the form that defaults to a
freshly instantiated managed user if not set.

* Impersonation form now includes the target user

So we no longer need to pass it around.

* Remove unused attr_reader

* Parameter dance to avoid hack

* Extract a method

* Merge method guards

* Stop saving authorizations for impersonations

We only need to check that the form is valid, really. It also makes more
since since these authorizations are ephimeral (they need to be checked
every time), so why save them?

* Slightly more robust test

Maybe a bit more brittle but more likely to catch bugs I think.

* Make impersonate form readonly

By making sure we feed it a user, and validate the persence of it, we no
longer need to mutate the form after creating it. This also makes both
impersonation & managed users controller more consistent with each
other.

* Get rid of the CreateManagedUser command

Merge it into the ImpersonateUser one since the logic is almost the same
now.

* Rename tests to better represent new situation

* Normalize impersonate form filling

* Test more realistic user flows

After all the refactoring, managed users do not have authorizations
stored in DB. So it makes more sense to test these cases for regular
users.

* Add reason to regular user impersonations

* Remove unnecessary line

We're already there at that point.

* Normalize markup

* Form switching without AJAX

Otherwise there's an annoying delay when switching forms, and it's not
necessary.

* Fix bad spec

* Merge both controllers into one

The logic is the same, so it makes sense to unify them so it's easier to
see how the whole thing works and what the edge cases are.

* Locale normalization after all the dance

* Unnest impersonations controller now

Since it doesn't apply only to managed users.

* Remove shared spec

To unlock easily adding further specs.

* Remove unnecessary nesting

* Simplify access to handler name in view

* Spare the select box when there's only one method

* Fill in missing specs for handler switching

* Consolidate abilities

* Move impersonation list to separate controller

The functionality is separate from the impersonations stuff, and that
controller is already complicated enough. By splitting it, we can also
choose a more adequate name for the resource.

* Indentation & whitespace

* Fill in missing bits of functionality

I forgot to add this when I worked on it at the very beginning of this
branch.

* Improve UI labelling to clarify functionality

* Generalize user filter query

Let it receive and arbitrary scope and not limit it to organization
users.

* Disallow impersonating admins

Not sure about the specs for this, but I'll stay on the safe side.

* Fix broken unit test

* Another pass of locale cleaning

* Rename test objects to reduce conflict chances

For example the default action on the dummy component is called "Foo" so
capybara actions are likely to conflict if we use that name.

* Fix last minute regression that I found

* Add change log entries

* Revert "Other dummy"

This reverts commit fb327d6cb660af44cb6b9947832dff2914ae4dc8.

I don't think we need this in the default application.

Specs were actually relying on the existance of the global extra
authorization handler, so they needed a bit of tweaking.