Fix gem manager to not always capture output
For example, when running `bundle exec rake test_admin`, the whole test
output would be captured, and spit at the end of the run. That's not
what we want. Capturing is only useful when we run generator tests to
swallow the really verbose output of the generator, and only spit it if
the test fail.