Use a better environment variable
I think this variable was previously unused (grepped for it in Rails
sources and found nothing). However, a modified form of it could be
useful for setting the root for code coverage, which was being
incorrectly set for external plugins. This is because external plugins
don't run tests from a subfolder but from the root of the repo so the
previous config was setting the root folder for coverage to the parent
folder of the plugin, potentially overwriting your shiny new project
named "coverage" :warning: