Generate separate DBs for development and test apps (#3299)
* Move test specific rake tasks to the tests

This avoids calling nested subshelling which makes it very hard to
actually print output for debugging. Also makes the decidim-generators
Rakefile "standard", which also feels nice.

* Generate separate DBs for development & test apps

Up to know, `development_app` and `spec/decidim_dummy_app` would share
their DBs, so regenerating the test app would drop your development app
DB (or would fail if your development application is running), and
viceverse. These are separate applications so they really shouldn't
interfere with each other.

This patch also generate different DB names for external plugins since
the base name for the app is inferred from the root folder name, so it
makes a previous fix to configure the test DB and development DB names
separately unnecessary.

* Fix gem manager to not always capture output

For example, when running `bundle exec rake test_admin`, the whole test
output would be captured, and spit at the end of the run. That's not
what we want. Capturing is only useful when we run generator tests to
swallow the really verbose output of the generator, and only spit it if
the test fail.

* Normalize LOAD_PATH manipulation

* Use a better environment variable

I think this variable was previously unused (grepped for it in Rails
sources and found nothing). However, a modified form of it could be
useful for setting the root for code coverage, which was being
incorrectly set for external plugins. This is because external plugins
don't run tests from a subfolder but from the root of the repo so the
previous config was setting the root folder for coverage to the parent
folder of the plugin, potentially overwriting your shiny new project
named "coverage" :warning:

* Remove "incorrect" filter

.rake files are regular ruby code, there's no reason to not tracking
coverage on them... If we are lazy for not writing tests for those
tests, simplecov should shout at us! :)

* Reuse `repo_root` method

* Lint .simplecov file

Since it just a (admittedly non-standard) ruby file.

* Fix a couple of BUNDLE_GEMFILE resolutions

* Remove unused file

* Move simplecov config to `.simplecov`

We don't need the no_defaults/defaults dance since to coverage root is
configurable directly before filters in the `SimpleCov.start` block.

By doing this, `require "simplecov"` behaves consistently from anywhere,
automatically applying our custom config.

* Make sure coverage tracking plays with generators

* Fix subprocess coverage merging by giving each subprocess a unique
command name.

* Make sure generator files are properly tracked by requring simplecov
early via RUBYOPT and setting the proper environment around the decidim
executable.

* Track coverage on generator specs

* s/restore_cache/restore_ruby_cache/

Since we restore other caches in this config as well.

* Extract a `save_ruby_cache` reference

* Send generators coverage to codeclimate

* Simplecov should never track js files...

* Fix external test app generation

Otherwise one gets

```
Invalid application name decidim_barcelona, constant DecidimBarcelona is
already in use. Please choose another application name.
```

* Remove {DEV,TEST}_DATABASE_NAME env variables

They were actually responsible for the issue of test & development DBs
overwriting each other! We no longer need them since now the generated
names take into account the base folder of the plugin where we're
working.