Fix permissions for space areas (#3401)
Once you entered a space area and selected a space instance (ie you
start managing a single initiative), the `current_participatory_space`
helper method is set to that initiative. This caused all the space
permissions to fail the check to enter the space area, as it's using
that value to check for the space type. Since, in those permissions
classes, the check of the participatory space is done before checking
the access to the space area, this caused the permission check to fail.

The solution was to move the space area check before the participatory
space type check.