Upgrade Rails to 5.2 (#3225)
* Prevent Rails 5.2 deprecation warnings

Like the following, or similar:

```
DEPRECATION WARNING: Dangerous query method (method whose arguments are
used as raw SQL) called with non-attribute argument(s): "name->'en'
ASC". Non-attribute arguments will be disallowed in Rails 6.0. This
method should not be called with user-provided values, such as request
parameters or model attributes. Known-safe values can be passed by
wrapping them in Arel.sql(). (called from index at
/path/to/decidim-admin/app/controllers/decidim/admin/scopes_controller.rb:13)
```

* Only install decidim migrations

This is necessary to ignore activestorage ones on Rails 5.2. This is not
obscure, it's "lo siguiente" :(. The answer is buried deep down in
activerecord's sources:

https://github.com/rails/rails/blob/662ba236d115d3e2152b04dcdefdc0ee6f1f6102/activerecord/lib/active_record/railties/databases.rake#L390

* Tweak bootsnap to play nice with byebug

* Bump Rails to 5.2

* Bootsnap is require in `config/boot.rb` of the default generated
application, so we need to add it to the Gemfile.

* Surveys were calling `save!` on `destroy!`'ed records, that's
(understandably) no longer valid and was only working by chance.

* A transaction block inside `Decidim.traceable` was no longer returning
its last statement, so I needed to add a explicit return. I also removed
a nested transaction block that seemed unnecessary.

* Upgraded papertrail and fixed a deprecation warning about `whodunnit`.

* Oauth's request env no longer has indifferent access. Apparently.

* Try fix hang by not mixin ruby & rails autoloading

* Try fix flaky in system UI