Fix nokogiri flakies (#3966)
* Remove barely used locale parameter

This parameter to the `have_i18n_content`/`have_no_i18n_content`
translation helpers was only used once, and with the default locale.
Remove it as part of a bigger refactor.

* Remove unnecessary `stripped`'s

* Simplify stripped implementation

So that it does not use Nokogiri.

* Fix incorrect linked resources spec

They were passing by chance due to previous implementation of the
`stripped` helper. It would accept a `nil` input a return the empty
string, which can be found on any page.

* Fix bad usage of `have_i18n_content`

It was passing by chance due to previous implementation of the
`stripped` helper. It would accept a `nil` input a return the empty
string, which can be found on any page.