Improved hashtag renderer performance
Before it was doing a query for each tag in the content. With these changes is performing only one query to retrieve all the hashtags. On the other hand, we are losing the GlobalId flexibility, but we were searching specifically for the Decidim::Hashstag GlobalId in the content, so it wasn't so flexible.