Allow cancelling a comment vote (#5042)
#### :tophat: What? Why?
Allows cancelling a comment vote (up or down) in case the user has already voted the comment. This is important regarding the UX to prevent users from locking down their opinion due to misclicks and producing unnecessary visibility (positive or negative) for the comment.

This is done by allowing the user to to re-click the voting button after the voting. Clicking it for a second time cancels the vote.

#### :pushpin: Related Issues
- Fixes #5033

#### :clipboard: Subtasks
- [x] Add `CHANGELOG` entry
- [X] Add tests