Improve the dynamic fields component for IE11 support
Because IE11 does not support the `<template>` tag very well,
add functionality to the dynamic fields component to match also
`<script type="text/template">` tags as an alternative.

The views need to use this type of templates in order for them to
work properly on IE11. Also, this types of templates cannot have
child templates meaning that the wrapper element needs to map the
child template with the `data-template` attribute.