Change the dynamic fields markup to be compatible with IE11
With the updated dynamic fields component, it is now possible to
mark the dynamic fields markup with
`<script type="text/template">`. This makes them compatible with
IE11.

Also note that this types of templates cannot have any child
templates which is why the child templates have been moved outside
of the parent template and mapped with the `data-template`
attribute on the dynamic fields wrapper element.