Fix dynamic fields elements for IE11 (#5052)
#### :tophat: What? Why?
This fixes the dynamic field elements for IE11 as explained in #5035.

This is an unfortunately complex fix for IE11 but it is the only way to make it compatible with that browser. The biggest issue with this is not actually the admin panel where we initially noticed issues with this but another component we built for the participation interface using the same component. Therefore, this component should consider both cases to support IE11 as well.

This should also maintain backwards compatibility in case someone is using the `<template>` tags with the dynamic fields component.

#### :pushpin: Related Issues
- Fixes #5035

#### :clipboard: Subtasks
- [x] Add `CHANGELOG` entry
- [X] Fix tests