Test for available locales (#5080)
#### :tophat: What? Why?

Regarding the fact that some files in the code need their localization version present or Rails crashes (as explained in #5079).

In this test I check the presence of a localized version of a datepicker file for every language available in the platform. The problem I faced is that these locales are defined in the module `Decidim` which is configured by the test suite and their default locales changed.

So my solution is a little "hacky" and reevaluates the module inside another to isolate it from the already instantiated.

Please let me know what you think. thanks!

#### :pushpin: Related Issues
- Related to #5079