Fix spec
I refactored a spec in the last commit but realized after fixing a rubocop
warning that the error was not being raised anymore.

The error is a regression that was fixed in #5179, which updated the component
setting :participatory_texts_enabled from true to false when the checkbox
was being disabled from the frontend.

I disabled the rubocop warning and made sure the error is catched now, testing
it in the branch that introduced the regression.