Set specific names for scopes.
After debugging this for too many hours, I've concluded that if a scope
has an "s" in its name, for some reason I fail to understand, it is
escaped and represented as &#115 (which is an "s" for html entitites).
This seems to happen at browser level and I can't reproduce it otuside
the test environment.