Update dependencies (#5206)
* Update dependencies

* Fix Faker deprecations

* Fix spec

* Add missing migration from Doorkeeper

* Always use latest Decidim Docker image

* Fix Rubocop issues

* Use Ruby grep

* Set specific names for scopes.

After debugging this for too many hours, I've concluded that if a scope
has an "s" in its name, for some reason I fail to understand, it is
escaped and represented as &#115 (which is an "s" for html entitites).
This seems to happen at browser level and I can't reproduce it otuside
the test environment.

* Remove deprecated PaperTrail config

* Don't run specs that depend on master when not on master