Fix managed users stolen identities with users having the same name
The managed users functionality should not consider the person's
name as a unique identifier for the person as there can be multiple
managed users with the same name.

This fixes the issue by creating new managed users every time in
case the same managed user with the same identity has not been
identifier before with the same authorization and identity.

See:
https://meta.decidim.org/processes/bug-report/f/210/proposals/14764