Listen on all interfaces in Docker
When starting decidim using the Docker setup, the server listens on 127.0.0.1 by default. This leads to connection reset errors when trying to access the application from outside the container. To fix this, tell the rails server to listen on all network interfaces (0.0.0.0) instead of just the loopback interface (127.0.0.1) when running in Docker.