Add a way to not filter by a field, but having all checkboxes checked
Having "" as default search criteria for a field do the same query that using "all", but checkboxes start unchecked, and this could be confusing.
Having an array with all possible values as default will check all checkboxes, but add extra conditions to the query. Sometimes these conditions leads to strange scenarios, since is not exactly the same filtering by an array with "all possible values" than not filtering.