Request confirmation to exit budgets component (#5765)
* Warn when trying to leave budgets component before completing the vote

* Add test for the alert before exiting the budgets component

* fix: base URL used to detect component URLs was including locale

* Component base URL based on current request instead of EngineRouter

It was working OK when using EngineRouter, but it was very complicated to get the tests passing. EngineRouter uses organization's host attribute, that can't store a port in tests, while the browser adds the host and port to the link's `href` attribute. On the other hand, Capybara uses random ports for the tested application, and #2187 adds a `base` tag to the pages to improve the debugging of failed tests HTML screenshots.

* docs: changelog entry added