Fix weird behavior when adding questions and moving them up
they were being added to the end of the element, but there are some invisible elements there, so moving the question up seemed not to work for a while, but it was actually moving betwenn invisible elements. This is now fixed, since new questions are added right after the last one, thus avoiding the invisible elements.