Fix the edit link test failing seemingly randomly previously
The assemblies spec does multiple requests to the target pages
when testing with the assemblies spec. This was caused by warden
failing to authenticate the user on the correct request because
in certain situations the previous request's rendering caused
other requests to "use" the authenticated state before the edit
link test did its own request.