Fix development_app rake task
Adding the profiling gems to the Gemfile midway through the install
generator breaks the next steps because they are still missing. That's
because we don't run `bundle install` in the generator. Moving this to the end makes it work.

This got broken in https://github.com/decidim/decidim/pull/6281. Thanks
to @microstudi for the heads-up!