Simplify file upload pre-conditions management
This generalizes most of the file upload management under
centralized configuration concern and also maps the file upload
conditions to the models where the uploaders are attached to.

This also introduces a new `PassthruValidator` record validator
that can be applied to the forms passing through the file upload
validations down to the model.