Toggle machine translations from URL
We need to use RequestStore because we translate things from inside presenters, which don't have any access to the request status and instance variables. With RequestStore we can store info per request. It's a dependeency of another dependency (PaperTrial), so this commit is just promoting it as a base dependency.