Change default initializer settings (#6566)
* Change domain.org and your-provider.org to example.org

"""
As described in RFC 2606 and RFC 6761, a number of domains
such as example.com and example.org are maintained for
documentation purposes. These domains may be used as
illustrative examples in documents without prior coordination
with us. They are not available for registration or transfer.
"""

See https://www.iana.org/domains/reserved

* Move locales settings near in default Initializer

* Remove Etherpad duplicated config in default Initializer

* Move 'machine translations' settings near in default Initializer