Do not modify the controller class in the controller tests that render views (#7755)
* Do not modify the controller class in the shared context

* Remove unnecessary `clear_helpers` call in the shared example

* Modify the view_context_class instead

If we override the `view_context` method on the controller, the
view assignment variables won't be available because the view
context won't be re-created after the view action.

This ensures the view assignments are available while keeping the
controller class untouched.

* Remove the unnecessary allow and_return

The view_controller_class will be stored in a class variable once
it has been called for the first time. Therefore, the allow
and_return is unnecessary.