Fix pt-BR issue
Helping on https://github.com/decidim/decidim/pull/6221#issuecomment-940410700 pt-BR issue. 
My guess is that with the current config: 
```
        pt-BR: pt-BR
        pt-PT: pt
```
Crowdin tries to have a 2 letters fallback, and as it is not provided, will take the first option (here pt-BR), ending having `pt-BR` labeled with `pt:`

So as in french, I propose to change it to have `pt: pt` to make sure default `pt-BR` is considered as a variant.