Revert the filter variable back to the params based proxy class
The filter class will not work for controllers/views which do not
define a dedicated search_collection such as the global search
page. For this reason, we revert back to the params based proxy
class for building the forms.

The `Ransack::Search` classes work just as well with the forms but
they need to have a searchable collection behind them.