Standardize the search scope names
Use the `with_*` search scopes everywhere where a search scope is
referring to a potential column name in the database.

This is not always the case, e.g. with the `with_date` or
`with_any_date` scopes which are targeting the start/end dates of
the records but for consistency, use the same naming convention
for all these scopes that are used within the ransack searches.