Rename shared contexts for following the same suffix as others
This is done just for a cosmetic change, to be consistent with the rest
of shared contexts that we have. In any of these cases it works the same
with our current RSpec configuration, the only requirement is to be in
the spec/shared/ folder.

I could rename the ones with *_shared_context.rb to just *_context.rb, as that
makes more sense IMHO: they're already in a spec/shared/ directory after
all, so having spec/shared/*_shared_context.rb seems a bit redundant. But it
seems that our convention is to have this *_shared_context.rb in the filename,
as there are more already with that naming (15 with _shared_context.rb vs 4
with _context.rb).