Fix the domain and flags in the default consent cookie
After introducing the new flags and setting the domain for the
cookie, the Capybara cookie was overriding the cookie set by JS
because it had incorrect domain (JS allows also subdomains).

Also add the extra flags to the default cookie that were added to
the JS for consistency and to avoid potential side effects.