Fix account update without password change (#9582)
* Handle password change properly at the account form

When the user tried to update their account without changing their
password, they could not submit the form due to the front-end
validations.

Fix this issue by marking those fields required only when they
are visible on the form.

* Make sure the spec checks the encrypted password is not changed

When account is updated without providing a password, the
encrypted password shouldn't change.

* Do not autofill the password on the account form

In tests, the user record responds to `.password` which causes
the account form updates to work differently in tests than they do
for actual users.

* Apply feedback from the code review

* Add also password and password confirmation errors on the form

In case the password update failed, the user doesn't get any
feedback. This fixes that issue.