Fix cryptic file validation errors
* Fix cryptic file validation errors

* Improve the content type validator and add method documentation

* Add content type validator spec for denylist case

* Fix the file definition within the spec

* Fix specs broken because of the message changes

* Update the error message for invalid file content type as per review

* Fix the obfuscated error messages also for file content type validators

Previously this only applied to the validators run for files
attached through the uploaders. Which is problematic e.g. for CSV
files that we just upload and process without an uploader
in-between.

* Add spec for the customized file content type validator

* Get the extension allowlist through the uploader if available

* Fix the extension_allowlist translation

* Move the extension allowlist translation to the correct module

* Fetch the extension allowlist through the correct recrod

Previously it was always using the defaults instead of the
settings configured for the target organization attached to the
record.

* Pass the context to the form records assigned to PassthruValidator

The votings census file upload broke after modifying the file
validation logic as it was missing the current participatory space
where it fetches the organization from.

* Pass the correct upload context for the dynamic file validations

In order to fetch the correct file extensions in each situation
for the dynamic upload errors, we need to define a proxy record
to pass the correct upload context for the validations.

* Sort the file extensions at the FileValidatorHumanizer

In order to show the file extensions similarly in the help texts
and the errors, sort them at the humanizer which produces the
extensions list for the help texts.

* Fix file not attached to organization with attachment proxy

* Normalize locales

* Use delegate method instead of extending SimpleDelegator