Fix trailing slash when using node.href
As the comment says, we use the href attribute (`node.getAttribute("href")`)
instead of the `node.href` property because the latter returns the URL with
a trailing slash, which is not what we want.
Before we were using jQuery for this (`$link.attr("href")`), so this
wasn't an issue, as it returned the URL withouth the trailing slash.