Ignore social share button links in external domain warning (#10121)
* Improve security of social share button links

* Fix typo

* Add updateExternalDomainLinks jest test

* Refactor 'exclude ancestor classes' feature

* Ignore social share button class in external domain warning

* Add missing semicolon

* Open social share links in a new tab

* Disable external link icon for social share links

* Undo EXCLUDE_CLASSES introduction in updateExternalDomainLinks

* Introduce ExternalDomainLink class so it's consistent with ExternalLink

Convert the old updateExternalDomainLinks function to the ExternalDomainLink class.
Ditch jQuery in favor of plain JavaScript.
The main idea behind this change is to make it behave more like the ExternalLink:
* same structure
* same logic of exclusion (with the `data-xx = false` attribute)

* Document exceptions for ExternalDomainLink

* Remove leftover class from EXCLUDE_CLASSES introduction

* Fix linter offenses

* Implement external domain link exclusion for SocialShare buttons

* Update ExternalDomainLink usage in comments

* Fix trailing slash when using node.href

As the comment says, we use the href attribute (`node.getAttribute("href")`)
instead of the `node.href` property because the latter returns the URL with
a trailing slash, which is not what we want.
Before we were using jQuery for this (`$link.attr("href")`), so this
wasn't an issue, as it returned the URL withouth the trailing slash.