Move routes definitions for decidim-templates to decidim-admin
If we don't do this, it's not respecting the decidim-admin constraints
that are already defined.

This was detected with a redirect when there was a normal user. Instead
of giving a 404 in /admin/templates/questionnaire_templates, it was
giving a redirection (301) to /admin. Moving the definition of the
routes to the decidim-admin module honors all the contracts from
the admin.